+++ docs/CHANGELOG.md
@@ -15,6 +15,8 @@ _Latest build from master branch._
   ([#121](https://github.com/Tubeshade/Tubeshade/issues/121))
 - Access control rules for new users and system user
   ([#264](https://github.com/Tubeshade/Tubeshade/pull/264))
+- Don't automatically reindex videos that have not yet been published
+  ([#265](https://github.com/Tubeshade/Tubeshade/issues/265))
 
 ## [0.1.5] - 2026-04-04